refatorando serviços REST

+++ src/main/java/com/algaworks/algafood/domain/service/CadastroCidadeService.java
@@ -14,6 +14,8 @@ import org.springframework.stereotype.Service;
 @Service
 public class CadastroCidadeService {
 
+  public static final String MSG_CIDADE_NAO_ENCONTRADA = "Não existe cadastro de estado com código %d";
+  public static final String MSG_CIDADE_EM_USO = "Cidade de código %d não pode ser removida, pois está em uso";
   @Autowired
   private CidadeRepository cidadeRepository;
 
@@ -26,7 +28,7 @@ public class CadastroCidadeService {
 
     Estado estado = estadoRepository.findById(estadoId).orElseThrow(
             () -> new EntidadeNaoEncontradaException(
-                    String.format("Não existe cadastro de estado com código %d", estadoId)
+                    String.format(MSG_CIDADE_NAO_ENCONTRADA, estadoId)
             ));
 
     cidade.setEstado(estado);
@@ -35,19 +37,31 @@ public class CadastroCidadeService {
 
   public void excluir(Long cidadeId) {
 
+    if (!cidadeRepository.existsById(cidadeId)) {
+      throw new EntidadeNaoEncontradaException(
+              String.format(MSG_CIDADE_NAO_ENCONTRADA, cidadeId));
+    }
+
     try {
       cidadeRepository.deleteById(cidadeId);
 
     } catch (EmptyResultDataAccessException e) {
       throw new EntidadeNaoEncontradaException(
-              String.format("Não existe um cadastro de cidade com código %d", cidadeId)
+              String.format(MSG_CIDADE_NAO_ENCONTRADA, cidadeId)
       );
     } catch (DataIntegrityViolationException e) {
       throw new EntidadeEmUsoException(
-              String.format("Cidade de código %d não pode ser removida, pois está em uso", cidadeId)
+              String.format(MSG_CIDADE_EM_USO, cidadeId)
       );
     }
 
+  }
+
+  public Cidade buscarOuFalhar(Long cidadeId) {
+    return cidadeRepository.findById(cidadeId)
+            .orElseThrow(() -> new EntidadeNaoEncontradaException(
+                    String.format(MSG_CIDADE_NAO_ENCONTRADA, cidadeId)
+            ));
   }
 
 }
